+++ bin/citrea/tests/evm/subscription.rs
@@ -6,6 +6,7 @@ use alloy_primitives::FixedBytes;
 use alloy_sol_types::SolEvent;
 use citrea_evm::smart_contracts::{AnotherLogEvent, LogEvent, LogsContract, TestContract};
 use citrea_evm::{Filter, LogResponse};
+use citrea_sequencer::SequencerConfig;
 // use citrea::initialize_logging;
 use citrea_stf::genesis_config::GenesisPaths;
 use reth_primitives::{keccak256, Address};
@@ -13,10 +14,11 @@ use tokio::time::sleep;
 
 use crate::evm::make_test_client;
 use crate::test_client::TestClient;
-use crate::test_helpers::{start_rollup, tempdir_with_children, wait_for_l2_block, NodeMode};
+use crate::test_helpers::{
+    create_default_rollup_config, start_rollup, tempdir_with_children, wait_for_l2_block, NodeMode,
+};
 use crate::{
-    DEFAULT_DEPOSIT_MEMPOOL_FETCH_LIMIT, DEFAULT_MIN_SOFT_CONFIRMATIONS_PER_COMMITMENT,
-    TEST_DATA_GENESIS_PATH,
+    TEST_DATA_GENESIS_PATH, TEST_SEND_NO_COMMITMENT_MIN_SOFT_CONFIRMATIONS_PER_COMMITMENT,
 };
 
 #[tokio::test(flavor = "multi_thread")]
@@ -26,22 +28,22 @@ async fn test_eth_subscriptions() -> Result<(), Box<dyn std::error::Error>> {
     let sequencer_db_dir = storage_dir.path().join("sequencer").to_path_buf();
 
     let (port_tx, port_rx) = tokio::sync::oneshot::channel();
-    let da_db_dir_cloned = da_db_dir.clone();
+
+    let rollup_config =
+        create_default_rollup_config(true, &sequencer_db_dir, &da_db_dir, NodeMode::SequencerNode);
+    let sequencer_config = SequencerConfig {
+        min_soft_confirmations_per_commitment:
+            TEST_SEND_NO_COMMITMENT_MIN_SOFT_CONFIRMATIONS_PER_COMMITMENT,
+        ..Default::default()
+    };
     let seq_task = tokio::spawn(async {
         // Don't provide a prover since the EVM is not currently provable
         start_rollup(
             port_tx,
             GenesisPaths::from_dir(TEST_DATA_GENESIS_PATH),
             None,
-            NodeMode::SequencerNode,
-            sequencer_db_dir,
-            da_db_dir_cloned,
-            DEFAULT_MIN_SOFT_CONFIRMATIONS_PER_COMMITMENT,
-            true,
-            None,
-            None,
-            Some(true),
-            DEFAULT_DEPOSIT_MEMPOOL_FETCH_LIMIT,
+            rollup_config,
+            Some(sequencer_config),
         )
         .await;
     });