+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## v0.5.2 (2024-09-30)
+- Added config for disableing prover proving session recovery. ([#1241](https://github.com/chainwayxyz/citrea/pull/1241))
+- Nodes now log each RPC request and response. ([#1236](https://github.com/chainwayxyz/citrea/pull/1236))
+
 ## v0.5.1 (2024-09-26)
 
 - Fix bug where full nodes would query more soft confirmations than intended. ([#1230](https://github.com/chainwayxyz/citrea/pull/1230))